# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for user management",
        "title": "User API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/services": {
            "get": {
                "description": "Get all Services",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Get All Services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Services.Service"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Create a Service",
                "parameters": [
                    {
                        "description": "Service object",
                        "name": "service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Services.Service"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            },
            "delete": {
                "description": "Remove all Services from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Remove all Services",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/services/search": {
            "get": {
                "description": "Search services",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Search services",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Services.Service"
                            }
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "description": "Get a Service by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Get a Service by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Services.Service"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a Service by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Services"
                ],
                "summary": "Update a Service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service object",
                        "name": "service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Services.Service"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete a Service by ID",
                "tags": [
                    "Services"
                ],
                "summary": "Delete a Service",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Services.Service": {
            "type": "object",
            "required": [
                "services_assets",
                "services_body",
                "services_title",
                "services_type"
            ],
            "properties": {
                "services_assets": {
                    "type": "string",
                    "example": "http://example.com/assets/service.jpg"
                },
                "services_body": {
                    "type": "string",
                    "example": "We specialize in building responsive and user-friendly websites."
                },
                "services_id": {
                    "type": "string"
                },
                "services_title": {
                    "type": "string",
                    "example": "Web Development"
                },
                "services_type": {
                    "type": "string",
                    "example": "Development"
                }
            }
        },
...
    }
}